pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["jkulds <w84my5@yandex.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
psycopg2 = "^2.9.9"
django-filter = "^23.5"
rollbar = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
